[#14] fix: 패키기 구조 변경
