--- src/main/java/net/mediger/api/auth/api/dto/RequestLogin.java
@@ -1,7 +1,0 @@
-package net.mediger.api.auth.api.dto;
-
-public record RequestLogin(
-        String account,
-        String password
-) {
-}
