--- src/main/java/net/mediger/api/auth/jwt/JwtFilter.java
@@ -1,65 +1,0 @@
-package net.mediger.api.auth.jwt;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.List;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Component
-@RequiredArgsConstructor
-public class JwtFilter extends OncePerRequestFilter {
-
-    private static final String AUTHORIZATION_HEADER = "Authorization";
-    private static final String TOKEN_PREFIX = "Bearer ";
-
-    private final TokenProvider tokenProvider;
-
-    @Override
-    protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response,
-                                    @NonNull FilterChain filterChain)
-            throws ServletException, IOException {
-        String token = resolveToken(request);
-
-        if (StringUtils.hasText(token) && isValidateToken(token)) {
-            setAuthentication(token);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String token = request.getHeader(AUTHORIZATION_HEADER);
-
-        if (StringUtils.hasText(token) && token.startsWith(TOKEN_PREFIX)) {
-            return token.substring(TOKEN_PREFIX.length());
-        }
-
-        return null;
-    }
-
-    private boolean isValidateToken(String token) {
-        return tokenProvider.validateToken(token);
-    }
-
-    private void setAuthentication(String token) {
-        String id = tokenProvider.getIdFromToken(token);
-        String role = tokenProvider.getRoleFromToken(token);
-        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(id, null,
-                getAuthority(role));
-        SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-    }
-
-    private List<SimpleGrantedAuthority> getAuthority(String role) {
-        return List.of(new SimpleGrantedAuthority(role));
-    }
-}
