--- src/main/java/net/mediger/api/auth/jwt/ResponseToken.java
@@ -1,9 +1,0 @@
-package net.mediger.api.auth.jwt;
-
-import lombok.Builder;
-
-@Builder
-public record ResponseToken(
-        String accessToken
-) {
-}
