--- src/main/java/net/mediger/api/auth/jwt/redis/RedisRepository.java
@@ -1,8 +1,0 @@
-package net.mediger.api.auth.jwt.redis;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RedisRepository extends CrudRepository<RefreshToken, String> {
-}
