--- src/main/java/net/mediger/api/member/api/dto/RequestBusinessDetails.java
@@ -1,9 +1,0 @@
-package net.mediger.api.member.api.dto;
-
-public record RequestBusinessDetails(
-        String address,
-        String ecommerceRegistrationNumber,
-        String settlementAccount,
-        String documents
-) {
-}
