--- src/main/java/net/mediger/api/member/domain/Role.java
@@ -1,5 +1,0 @@
-package net.mediger.api.member.domain;
-
-public enum Role {
-    MEMBER, BUSINESS, ADMIN
-}
