+++ src/main/java/net/mediger/auth/api/dto/RequestLogin.java
@@ -1,0 +1,7 @@
+package net.mediger.auth.api.dto;
+
+public record RequestLogin(
+        String account,
+        String password
+) {
+}
