+++ src/main/java/net/mediger/auth/jwt/ResponseToken.java
@@ -1,0 +1,9 @@
+package net.mediger.auth.jwt;
+
+import lombok.Builder;
+
+@Builder
+public record ResponseToken(
+        String accessToken
+) {
+}
