+++ src/main/java/net/mediger/member/api/dto/RequestBusinessDetails.java
@@ -1,0 +1,9 @@
+package net.mediger.member.api.dto;
+
+public record RequestBusinessDetails(
+        String address,
+        String ecommerceRegistrationNumber,
+        String settlementAccount,
+        String documents
+) {
+}
