+++ src/main/java/net/mediger/member/domain/Role.java
@@ -1,0 +1,5 @@
+package net.mediger.member.domain;
+
+public enum Role {
+    MEMBER, BUSINESS, ADMIN
+}
