+++ src/main/java/net/mediger/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package net.mediger.member.repository;
+
+import java.util.Optional;
+import net.mediger.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    boolean existsByAccount(String account);
+    Optional<Member> findMemberByAccount(String account);
+}
